check endiannes of a system

+++ 0x14-bit_manipulation/100-get_endianness.c
@@ -6,7 +6,7 @@
  */
 int get_endianness(void)
 {
-	short int i;
+	unsigned int i;
 	char *p;
 
 	i = 1;
